sys/dev: five defects from the first full model check of the driver tree
1,419 (file, function) pairs across 570 translation units and 189
sys/dev subdirectories -- the first time CBMC has covered the driver
tree at this width.

  PROVED   559     TIMEOUT  318
  BOUNDED   26     ERROR    100
  FAILED   416

THE ANSWER RATE IS THE HEADLINE CAVEAT AND BELONGS BESIDE THE PROOFS:
418 of 1,419 (29.5%) produced no verdict, and those functions are
unchecked rather than clean. 88 of the 100 ERRORs are the per-process
--mem-mb 4096 RLIMIT_AS firing on hard SAT instances, which is the
documented "one function unproved beats a shard lost" behaviour.

Of the 416 FAILED: 305 pointer/memory, 63 static-deferred, 24 EXPORTED
arithmetic, 10 address-taken static, 7 per-CPU, 5 INDEX, 2
extern-driven. 39 records in the three READ-THESE buckets, 29 of them
carrying a line not in the not-a-defect table. Five are defects.

rccgpio: the field is the mask, and six sites shift by it
---------------------------------------------------------
	{ .pin = (1 << 11), .name = "reset switch", ... },
	{ .pin = (1 << 15), .name = "red LED",      ... },
	{ .pin = (1 << 17), .name = "green LED",    ... },

attach() passes .pin straight to rcc_gpio_modify_bits(sc, reg, 0,
rcc_pins[i].pin) -- the third parameter is named `mask' and the body is
`value &= ~mask; value |= writebits'. Six other sites shift BY it:

	sc->sc_output |= (1 << rcc_pins[pin].pin);   /* 1 << 2048 */

1 << 2048, 1 << 32768, 1 << 131072. Undefined, and on x86 the count is
masked to five bits, so all three become 0: every pin collapses onto
bit 0 and THE DRIVER'S GPIO AND LED CONTROL CANNOT WORK AS WRITTEN.

THIS OVERTURNS A TRIAGE ALREADY IN THE DOCUMENT. It read: "`pin' is
bounded by the caller's GPIO_PIN_MAX check, one frame up." True and
irrelevant -- it bounds the INDEX, which all six entry points already
test. The defect is the value in the FIELD. The old sentence is struck
through rather than deleted, because a triage that was wrong has to stay
visible; and it is exactly the case triaged()'s docstring describes,
working: the claim sits in prose after the table, so the marker
correctly never said anybody had read it, and the record surfaced again.

arswitch: the read sibling bounds phy, the write one did not
------------------------------------------------------------
arswitch_readphy_internal bounds phy and reg; arswitch_writephy_internal
bounded only reg and then shifted by phy. phy arrives from
IOETHERSWITCHSETPHYREG through ETHERSWITCH_WRITEPHYREG(etherswitch,
phyreg->phy, ...) with no check of its own. Root-only (0600 root:wheel),
and every other etherswitch driver -- mtkswitch, adm6996fc, ukswitch,
rtl8366rb, e6060sw -- checks it on both sides. Same shape as the
es_vlangroup bound already fixed across all eight.

dmc620 and cmn600: register bounds the unit, unregister did not
---------------------------------------------------------------
Firmware declaring more than DMC620_UNIT_MAX (16) PMUs attaches those
units fine -- register silently skips them -- and detaching one writes a
NULL past dmc620_pmcs[] and decrements a count never incremented for it.
sys/arm64/arm64/cmn600.c is character-for-character the same against
cmn600_pmcs[4] and is fixed with it.

lge and sis: data << 16 is UB on every PHY reset
-------------------------------------------------
`data' is the int parameter of miibus_writereg carrying a 16-bit PHY
register value, and mii_phy_reset() writes BMCR_RESET == 0x8000, so
`data << 16' is 2^31. THE ORDINARY PHY-ATTACH PATH, not an edge case: a
UBSan kernel traps on every lge and sis attach. No functional
misbehaviour, and the correct idiom is in-tree already -- rt2860,
if_run and if_mtw all write (uint32_t)val << 16.

sis was NOT flagged by CBMC; it came out of surveying the idiom after
lge. Recorded separately so the provenance is weighable.

Two shift-into-bit-31 cases, a class fixed twice before
--------------------------------------------------------
DC_DEVID(vendor, device) is ((device) << 16 | (vendor)), and resolving
all 74 dc_devs[] entries against the header gives 30 with a device ID
>= 0x8000. Constants, folded correctly, so UB only; casting the macro
parameter fixes all 74 uses at once.

pci.c:6096 is the same shape on a DEVICE-SUPPLIED RUNTIME VALUE in core
PCI code: `cfg->device << 16 | cfg->vendor' with cfg->device a uint16_t
promoting to int. Gated on PCI_IOV + PCICFG_VF + PCIR_VENDOR + width 4,
the SR-IOV VF emulation, and PCI_IOV is in amd64 GENERIC so it ships. It
joins pci_ea_fill_info's three-bit index into dw[4] in the same marker
list -- the same class one function apart, which is why the duplicate-key
gate firing here was the right outcome rather than an obstacle.

Measured
--------
Analyse over every touched scope, before and after: 81 translation
units, 71 OK, 10 ERROR, 1 finding, identical both times and no per-file
change. All six edited .c files are at zero findings in both, and
cmn600.c analyses OK at zero. Eight markers added, all eight
revert-verified (exit 1 reverted, 0 restored). The duplicate-key gate
caught pci.c having an entry already; merged into a list rather than
shadowing it. sh tools/setup_local.sh --check is clean.

Three instrument findings from the run, in the document
--------------------------------------------------------
extern_driven() misses the commonest driver shape: it matches
`return_value_<callee>', but the recurring sys/dev case is a FIELD read
through an unmodelled device_get_softc, so the description says
`ch->unit * 2'. Several records landed in READ-THESE for that alone.

Two shared-header artefacts account for 6 of the 29 and will reappear in
every scope: systm.h:481's `tick_sbt * timo' inside pause(), and
time.h:145's `(sbintime_t)_bt.sec << 32' inside bttosbt(). Both are
unconstrained kernel globals; they may deserve a named rule beside the
per-CPU one.

report.py prints only the EXPORTED bucket under "worth a person's time",
so the INDEX and address-taken-static buckets -- where the rccgpio and
dmc620 findings live -- are counted and never listed, and its _all_read
check looks only at the first two failures. That is why this run's list
had to be dumped by hand.

+++ hbsd/src/sys/dev/dc/if_dcreg.h
@@ -1047,7 +1047,12 @@ struct dc_softc {
 #define	DC_DEVICEID_M5261	0x5261
 #define	DC_DEVICEID_M5263	0x5263
 
-#define	DC_DEVID(vendor, device)	((device) << 16 | (vendor))
+/*
+ * PBSD: (uint32_t)(device). 30 of the 74 dc_devs[] entries have a
+ * device ID >= 0x8000, and `(device) << 16' on a plain int is
+ * undefined. Casting the parameter fixes all 74 uses at once.
+ */
+#define	DC_DEVID(vendor, device)	((uint32_t)(device) << 16 | (vendor))
 
 /*
  * PCI low memory base and low I/O base register, and
